renamed helper_methods to utils
