+++ src/grader/autograder.py
@@ -6,7 +6,7 @@ import shutil
 from tempfile import TemporaryDirectory
 from typing import Tuple, List
 
-from helper_methods import *
+from grader.utils import *
 from test_suite import TestSuite
 from rubric import Rubric
